src/core/views/view_custom_user.py
from rest_framework.permissions import IsAuthenticated
from dj_rest_auth.views import UserDetailsView

from src.core.serializers.serializer_custom_user import CustomUserSerializer


class CustomUserDetailView(UserDetailsView):
    serializer_class = CustomUserSerializer
    permission_classes = [IsAuthenticated]

    def get_object(self):
        return self.request.user

    def list(self, request, *args, **kwargs):
        return super().list(request, *args, **kwargs)

    def retrieve(self, request, *args, **kwargs):
        return super().retrieve(request, *args, **kwargs)

    def update(self, request, *args, **kwargs):
        return super().update(request, *args, **kwargs)

    def destroy(self, request, *args, **kwargs):
        return super().destroy(request, *args, **kwargs)
