requirements.txt
asgiref==3.7.2
botocore==1.31.1
certifi==2024.2.2
dj-rest-auth==4.0.1
Django==4.2.10
django-cors-headers==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-stdimage==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
flake8==6.1.0
importlib-metadata==6.8.0
jmespath==1.0.1
Markdown==3.4.3
Pillow==10.4.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
zipp==3.15.0
googlemaps==4.10.0
google-cloud-tasks==2.14.2
mailersend==0.5.6
beautifulsoup4==4.12.3
webcrypt==1.3.1

gunicorn>=20.0,<21.0
dj-database-url==2.0.0
typing_extensions==4.9.0
psycopg2-binary
django-storages[google]
django-environ